SQLServer_AgenticAI/requirements.txt
langchain==1.2.0
langchain-openai==1.1.6
langchain-community==0.4.1
langgraph==1.0.5
sqlalchemy==2.0.45
pyodbc==5.3.0
ragas==0.4.2
datasets==4.4.2
openai==2.14.0
python-dotenv==1.2.1
streamlit==1.52.2
